Fix existing tests.
