+++ test/EFCore.MySql.FunctionalTests/Query/GearsOfWarQueryMySqlFixture.cs
@@ -1,4 +1,3 @@
-using System;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Pomelo.EntityFrameworkCore.MySql.FunctionalTests.TestUtilities;
@@ -35,22 +34,10 @@ namespace Pomelo.EntityFrameworkCore.MySql.FunctionalTests.Query
 
             foreach (var mission in data.Missions)
             {
-                mission.Timeline = GetExpectedValue(mission.Timeline);
+                mission.Timeline = MySqlTestHelpers.GetExpectedValue(mission.Timeline);
             }
 
             return data;
-        }
-
-        public static DateTimeOffset GetExpectedValue(DateTimeOffset value)
-        {
-            const int mySqlMaxMillisecondDecimalPlaces = 6;
-            var decimalPlacesFactor = (decimal)Math.Pow(10, 7 - mySqlMaxMillisecondDecimalPlaces);
-
-            // Change DateTimeOffset values, because MySQL does not preserve offsets and has a maximum of 6 decimal places, in contrast to
-            // .NET which has 7.
-            return new DateTimeOffset(
-                (long)(Math.Truncate(value.UtcTicks / decimalPlacesFactor) * decimalPlacesFactor),
-                TimeSpan.Zero);
         }
     }
 }
