+++ test/EFCore.MySql.FunctionalTests/Query/NorthwindIncludeQueryMySqlTest.cs
@@ -33,7 +33,7 @@ namespace Pomelo.EntityFrameworkCore.MySql.FunctionalTests.Query
 
         public override Task Include_collection_with_multiple_conditional_order_by(bool async)
         {
-            // The order of `Orders` can be different, becaues it is not explicitly sorted.
+            // The order of `Orders` can be different, because it is not explicitly sorted.
             // This is the case on MariaDB.
             return AssertQuery(
                 async,
@@ -49,7 +49,7 @@ namespace Pomelo.EntityFrameworkCore.MySql.FunctionalTests.Query
 
         public override Task Include_duplicate_collection_result_operator(bool async)
         {
-            // The order of `Orders` can be different, becaues it is not explicitly sorted.
+            // The order of `Orders` can be different, because it is not explicitly sorted.
             // This is the case on MariaDB.
             return AssertQuery(
                 async,
@@ -67,7 +67,7 @@ namespace Pomelo.EntityFrameworkCore.MySql.FunctionalTests.Query
 
         public override Task Include_duplicate_collection_result_operator2(bool async)
         {
-            // The order of `Orders` can be different, becaues it is not explicitly sorted.
+            // The order of `Orders` can be different, because it is not explicitly sorted.
             // This is the case on MariaDB.
             return AssertQuery(
                 async,
